fix: share implementation
